backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.26.0
PyMySQL==1.1.0
